move calculateMedian func to utils
