+++ x/oracle/abci/util.go
@@ -1,7 +1,10 @@
 package abci
 
 import (
+	"sort"
+
 	sdk "github.com/cosmos/cosmos-sdk/types"
+	oracletypes "github.com/ojo-network/ojo/x/oracle/types"
 )
 
 // VoteExtensionsEnabled determines if vote extensions are enabled for the current block.
@@ -20,4 +23,19 @@ func VoteExtensionsEnabled(ctx sdk.Context) bool {
 	}
 
 	return cp.Abci.VoteExtensionsEnableHeight < ctx.BlockHeight()
+}
+
+func calculateMedian(gasEstimates []oracletypes.GasEstimate) (median oracletypes.GasEstimate) {
+	sort.Slice(gasEstimates, func(i, j int) bool {
+		return gasEstimates[i].GasEstimation < gasEstimates[j].GasEstimation
+	})
+
+	mid := len(gasEstimates) / 2
+	if len(gasEstimates)%2 == 0 {
+		return oracletypes.GasEstimate{
+			GasEstimation: gasEstimates[mid-1].GasEstimation + gasEstimates[mid].GasEstimation,
+			Network:       gasEstimates[mid-1].Network,
+		}
+	}
+	return gasEstimates[mid]
 }
